requirements.txt
# --- Core & Data Processing ---
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
pyyaml
tqdm
joblib

# --- Deep Learning (PyTorch) ---
torch==2.2.2+cu118 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
torchvision
torchaudio

# --- NLP & Text Processing ---
emoji
pyvi
regex

# # --- Web App & API ---
streamlit==1.39.0
fastapi==0.115.0
uvicorn[standard]
python-multipart

# --- Visualization ---
matplotlib
seaborn
